requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.10
Werkzeug==0.9.6
argparse==1.2.1
coverage==3.7.1
docker-py==0.4.0
dockerpty==0.2.3
docopt==0.6.1
itsdangerous==0.24
mock==1.0.1
pygraphviz==1.2
requests==2.2.1
six==1.7.3
texttable==0.8.1
websocket-client==0.11.0
wsgiref==0.1.2
